+++ react-application-spring/src/main/java/com/ecommerce/react_application_spring/Repository/CustomersRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.ecommerce.react_application_spring.Model.Customers;
 @Repository
-public interface CustomersRepository extends JpaRepository<Customers,Integer> {
+public interface CustomersRepository extends JpaRepository<Customers,Long> {
     
 }
